Move to a mixin to be included by openshift
